+++ main.c
@@ -4,7 +4,6 @@ int	main(int argc, char **argv, char **envp)
 {
 	char		*input;
 	t_command	*command_list;
-	//t_data		data;
 	t_env	*env_list;
 
 	env_list = malloc(sizeof(t_env));
@@ -12,9 +11,7 @@ int	main(int argc, char **argv, char **envp)
 	(void)argc;
 	(void)argv;
 	command_list = NULL;
-	//data.env = envp;
 	create_env_list(envp, &env_list);
-	//printf_list(env_list);
 	while (1)
 	{
 		signal_handle();
@@ -27,9 +24,8 @@ int	main(int argc, char **argv, char **envp)
 			free(input);
 			return (1);
 		}
-		//print_commands(command_list);
 		free(input);
-		commands_manager(command_list, &env_list);
+		test_pipe(command_list, &env_list);
 		free_commands(command_list);
 		command_list = NULL;
 	}
